Migrate JUnit4 samples to JUnit 5
* Fixed imports in test templates
* Using subprojects{} to configure the test tasks for `junitPlatform`
