--- subprojects/docs/src/samples/java/components-with-junit-4-tests/groovy/build.gradle
@@ -1,8 +1,0 @@
-allprojects {
-    version = '1.0.2'
-    group = 'org.gradle.sample'
-
-    repositories {
-        jcenter()
-    }
-}
