--- subprojects/docs/src/samples/java/components-with-junit-4-tests/groovy/list/build.gradle
@@ -1,7 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-dependencies {
-    testImplementation 'junit:junit:4.12'
-}
