--- subprojects/docs/src/samples/java/components-with-junit-4-tests/groovy/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name = 'components-with-junit-4-tests'
-include 'application', 'list', 'utilities'
