--- subprojects/docs/src/samples/java/components-with-junit-4-tests/kotlin/application/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    application
-}
-
-dependencies {
-    implementation(project(":list"))
-    implementation(project(":utilities"))
-    testImplementation("junit:junit:4.12")
-}
-
-application {
-    mainClassName = "org.gradle.sample.Main"
-}
