--- subprojects/docs/src/samples/java/components-with-junit-4-tests/kotlin/list/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    `java-library`
-}
-
-dependencies {
-    testImplementation("junit:junit:4.12")
-}
