+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/groovy/application/build.gradle
@@ -1,0 +1,14 @@
+plugins {
+    id 'application'
+}
+
+dependencies {
+    implementation project(':list')
+    implementation project(':utilities')
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.1'
+    testRuntimeOnly 'org.junit.vintage:junit-vintage-engine'
+}
+
+application {
+    mainClassName = 'org.gradle.sample.Main'
+}
