+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/groovy/build.gradle
@@ -1,0 +1,16 @@
+allprojects {
+    version = '1.0.2'
+    group = 'org.gradle.sample'
+
+    repositories {
+        jcenter()
+    }
+}
+
+subprojects {
+    apply plugin: "java"
+
+    test {
+        useJUnitPlatform()
+    }
+}
