+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/groovy/list/build.gradle
