+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/kotlin/application/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    application
+}
+
+dependencies {
+    implementation(project(":list"))
+    implementation(project(":utilities"))
+    testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.0")
+    testRuntimeOnly("org.junit.vintage:junit-vintage-engine")
+}
+
+application {
+    mainClassName = "org.gradle.sample.Main"
+}
