+++ subprojects/docs/src/samples/java/components-with-junit-5-tests/kotlin/settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "components-with-junit-5-tests"
+include("application", "list", "utilities")
