+++ subprojects/docs/src/samples/jvm-components-with-additional-test-types/groovy/application/build.gradle
@@ -5,7 +5,8 @@ plugins {
 dependencies {
     implementation project(':list')
     implementation project(':utilities')
-    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.1'
+    testRuntimeOnly 'org.junit.vintage:junit-vintage-engine'
 }
 
 application {
@@ -25,7 +26,8 @@ configurations {
 }
 
 dependencies {
-    integrationTestImplementation 'junit:junit:4.12'
+    integrationTestImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.1'
+    integrationTestRuntimeOnly 'org.junit.vintage:junit-vintage-engine'
 }
 
 def integrationTest = tasks.register('integrationTest', Test) {
