+++ subprojects/docs/src/samples/jvm-components-with-additional-test-types/kotlin/utilities/build.gradle.kts
@@ -20,7 +20,8 @@ val integrationTestImplementation by configurations.getting {
 configurations["integrationTestRuntimeOnly"].extendsFrom(configurations.runtimeOnly.get())
 
 dependencies {
-    integrationTestImplementation("junit:junit:4.12")
+    integrationTestImplementation("org.junit.jupiter:junit-jupiter-api:5.6.1")
+    "integrationTestRuntimeOnly"("org.junit.vintage:junit-vintage-engine")
 }
 
 val integrationTest = tasks.register<Test>("integrationTest") {
