--- subprojects/docs/src/samples/templates/java-junit4-integration-test-for-utilities-library/src/integrationTest/java/org/gradle/sample/StringUtilsIntegrationTest.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.sample;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class StringUtilsIntegrationTest {
-    @Test public void testSplit() {
-        LinkedList list = StringUtils.split("The dog is green");
-        assertEquals(4, list.size());
-        assertEquals("The", list.get(0));
-        assertEquals("dog", list.get(1));
-        assertEquals("is", list.get(2));
-        assertEquals("green", list.get(3));
-    }
-}
